requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.1.0
Django==4.2.1
django-storages==1.14.2
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
gTTS==2.5.1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==23.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
djangorestframework==3.15.1